Fix prototype#map typing regression
The refactors to prototype#map in [#45][] are a little too presumptuous
in that they assume the integration optons values stored in metadata are
always going to be strings, which is not the case

In the long run, trying to determine if something is an array or a map
or a mixed on the client is a pretty bad solution. We're essentially
reduced to duck typing metadata structures, which is not a good
solution.

Instead metadata should look like:

```js
{
  "type": "mixed",
  "data": [
    { key: "whatever", value: "whatever_value" },
    // ...
  ]
}
```

[#45]: https://github.com/segmentio/analytics.js-integration/pull/45